Login: move existing code to a Fragment, MvRx style
